[SYCL][NFC] Refactor [[intel::num_simd_work_items()]] attribute tests
This patch updates tests for [[intel::num_simd_work_items()]] attribute:

1. separates AST and diagnostics to make it easier to read/follow-up.
2. removes duplicate test cases.
3. adds missing tests.
4. no compiler changes.
